+++ R/mirai-package.R
@@ -93,10 +93,12 @@ NULL
     n_zero = "the number of daemons must be zero or greater",
     numeric_n = "'n' must be numeric, did you mean to provide 'url'?",
     requires_list = "'.args' must be specified as a list",
-    sync_dispatch = "sync with dispatcher timed out after %s ms, perhaps specify a larger value?",
+    sync_dispatch = "initial sync with dispatcher timed out after 5s, specify 'dispatcher = NA' to allow longer",
+    sync_na = "initial sync with dispatcher timed out after 20s",
     sync_timeout = "sync with local process timed out after 5s",
     url_spec = "numeric value for 'url' is out of bounds"
   ),
   hash = TRUE
 )
 .timelimit <- 5000L
+.nalimit <- 20000L